refactor(op-devstack): move rust binary tooling to shared/rustbin, use it for kona-host
The shared/challenger package hardcoded kona-host to
rust/target/release/kona-host, bypassing the JIT build, env var
overrides, and debug/release discovery that other Rust binaries
already had via EnsureRustBinary in sysgo.

Move the rust binary resolution logic from sysgo into
shared/rustbin so it can be used by shared/challenger directly.
Decouple from devtest.CommonT by accepting context.Context and
log.Logger (nil-safe) instead. Update shared/challenger to resolve
kona-host through rustbin.EnsureExists, giving it the same
RUST_JIT_BUILD, RUST_BINARY_PATH_KONA_HOST, and multi-profile
search support that op-reth, kona-node, and other Rust binaries
already have.
